requirements.txt
flask 
flask-cors
google-generativeai 
python-dotenv 
requests 
ddgs
duckduckgo-search